fix: Issue #1
On Linux, the names of the header files must be case sensitive, but in the testclient.h files for the tests test_12_client and test_14_server there is: #include <Client. h>
This is incorrect because the Client.h file corresponds to that of the Arduino framework. You must put #include <client.h>